agents_catalog/standalone_agents/01-vehicle_to_grid/requirements.txt
fastmcp>=0.1.0
httpx>=0.24.0
boto3>=1.26.0
pydantic>=2.0.0
mcp>=0.1.0
strands-agents>=0.1.4
strands-agents-tools>=0.1.3
bedrock-agentcore>=0.1.1
bedrock-agentcore-starter-toolkit>=0.1.2
